- rename `testDeleteShow.py` to `testDelete.py` - set `delete_episode_of_show` to redirect to the show's index page instead of the request's referer
